Made code smaller by adding Thrift as maven dep, added removeFromWhitelist.
Added thrift as a maven dependency, this decreased the compiled jar size
from about 1MB to about 800KB as I was able to strip out all the
unnecessary thrift code from the codebase.

Added the removeFromWhitelist method, and changed the addToWhitelist method
to do nothing if the user is already whitelisted. The removeFromWhitelist
method also does nothing if the user is NOT on the whitelist.