Tweaked the console line class to include the level seperately.
The Log Level of each console line is included as a seperate property
of the ConsoleLine object now. This allows clients to filter and highlight
different levels of Log Message.

Also switched the reloadServer method from a oneway void to a bool, because
oneway methods do not play nice with C# (I can't figure out why not). I'm
sure that it's not a framework limitation and actually my implementation
of it that's causing the issues, but c'est la vie.